requirements.txt
bcrypt==4.0.1
cachetools==5.2.1
colorama==0.4.6
getch==1.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
oauthlib==3.2.2
pwinput==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.9
termcolor==2.2.0
